lab3/requirements.txt
gradio
upstash-vector
torch
torchvision
Pillow
transformers
python-dotenv
sentence-transformers
